+++ src/commands/actions/stats.js
@@ -1,4 +1,6 @@
 import Promise from 'bluebird';
+import randomColor from 'randomcolor';
+import converter from 'hex2dec';
 
 import { getUserAction } from '../../redis';
 
@@ -12,31 +14,19 @@ function userStatistics(client, evt, suffix) {
   getUserAction(userToCheck.id)
   .then(results => {
     if (!results) return Promise.resolve(evt.message.channel.sendMessage('', false, {color: 16763981, description: `\u26A0 No data found for this user :( - Go give them a hug!`}));
+    let hexCode = randomColor();
+    let cleanHex = hexCode.replace('#', '0x');
+    let embedColor = converter.hexToDec(cleanHex);
     let embed = {
-      color: 16737843,
+      color: embedColor,
       author: {
-        name: `Action statistics for ${userToCheck.username}#${userToCheck.discriminator}`
+        name: `Actions ${userToCheck.username}#${userToCheck.discriminator} received`
       },
-      fields: [
-        { name: 'Fruits Eaten:',
-          value: `\uD83C\uDF4E Apples: ${results.actions_foods_apple || 0} - \uD83C\uDF50 Pears: ${results.actions_foods_pear || 0} - \uD83C\uDF4A Tangerines: ${results.actions_foods_tangerine || 0}
-\uD83C\uDF4C Bananas: ${results.actions_foods_banana || 0} - \uD83C\uDF49 Watermelons: ${results.actions_foods_watermelon || 0} - \uD83C\uDF53 Strawberries: ${results.actions_foods_strawberry || 0}
-\uD83C\uDF51 Peaches: ${results.actions_foods_peach || 0} - \uD83C\uDF52 Cherries: ${results.actions_foods_cherry || 0} - \uD83C\uDF4D Pineapples: ${results.actions_foods_pineapple || 0}`,
-          inline: false },
-        { name: 'Meals Eaten:',
-          value: `\uD83C\uDF54 Hamburgers: ${results.actions_foods_hamburger || 0} - \uD83C\uDF5F Fries: ${results.actions_foods_fries || 0} - \uD83C\uDF2E Tacos: ${results.actions_foods_taco || 0}
-\uD83C\uDF63 Sushi: ${results.actions_foods_sushi || 0} - \uD83C\uDF5C Ramen: ${results.actions_foods_ramen || 0} - \uD83C\uDF72 Stew: ${results.actions_foods_stew || 0}
-\uD83C\uDF5D Spaghetti: ${results.actions_foods_spaghetti || 0} - \uD83C\uDF5B Curries: ${results.actions_foods_curry || 0} - \uD83C\uDF55 Pizzas: ${results.actions_foods_pizza || 0}`,
-          inline: false },
-        { name: 'Insects Eaten:',
-          value: `\uD83D\uDC1B Bugs: ${results.actions_foods_bug || 0} - \uD83D\uDC1E Beetles: ${results.actions_foods_beetle || 0} - \uD83D\uDD77 Spiders: ${results.actions_foods_spider || 0}`,
-          inline: false },
-        { name: 'Actions Received:',
-          value: `<:hug:457179730333007874> Hugs: ${results.actions_hugs || 0} - <:kiss:457179730349654018> Kisses: ${results.actions_kisses || 0} - <:boop:457182422962929676> Boops: ${results.actions_boops || 0}
-<:shoot:457179730525945856> Shots: ${results.actions_shots || 0} - <:dizzy:457184074336043028> Slaps: ${results.actions_slaps || 0} - <:pet:457178504442806292> Pets: ${results.actions_pets || 0}
-<:lick:457179730626478080> Licks: ${results.actions_licks || 0} - <:bite:436215483230846976> Bites: ${results.actions_bites || 0} - \ud83d\udc4b Spanks: ${results.actions_spanks || 0}`,
-          inline: false }
-      ]
+      description: `These are the actions you've received:\n
+      \uD83C\uDF4E **Fed:** ${results.actions_foods || 0}x - <:hug:457179730333007874> **Hugged:** ${results.actions_hugs || 0}x - <:kiss:457179730349654018> **Kissed:** ${results.actions_kisses || 0}x
+      <:boop:457182422962929676> **Booped:** ${results.actions_boops || 0}x - <:shoot:457179730525945856> **Shot:** ${results.actions_shots || 0}x - <:dizzy:457184074336043028> **Slapped:** ${results.actions_slaps || 0}x
+      <:pet:457178504442806292> **Pet:** ${results.actions_pets || 0}x - <:lick:457179730626478080> **Licked:** ${results.actions_licks || 0}x - <:bite:436215483230846976> **Bitten:** ${results.actions_bites || 0}x
+      \ud83d\udc4b **Spanked:** ${results.actions_spanks || 0}x - <:nuzzle:494097741203505152> **Nuzzled:** ${results.actions_nuzzles || 0}x - <:poke:494105957752897537> **Poked:** ${results.actions_pokes || 0}x`
     };
     return Promise.resolve(evt.message.channel.sendMessage('', false, embed));
   });
@@ -44,8 +34,7 @@ function userStatistics(client, evt, suffix) {
 }
 
 export default {
-  actions: userStatistics,
-  ustats: userStatistics
+  actioninfo: userStatistics
 };
 
 export const help = {